// src/core/modules/dao-api/models/Treasury.ts
/**
 * DAO-API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Token } from '../models/Token';
import { Transfer } from '../models/Transfer';
import { HttpFile } from '../http/http';

export class Treasury {
    'totalValue'?: any | null;
    'dailyChange'?: any | null;
    'dailyChangePercentage'?: any | null;
    'tokens'?: Array<Token>;
    'recentTransfers'?: Array<Transfer>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "totalValue",
            "baseName": "total_value",
            "type": "any",
            "format": ""
        },
        {
            "name": "dailyChange",
            "baseName": "daily_change",
            "type": "any",
            "format": ""
        },
        {
            "name": "dailyChangePercentage",
            "baseName": "daily_change_percentage",
            "type": "any",
            "format": ""
        },
        {
            "name": "tokens",
            "baseName": "tokens",
            "type": "Array<Token>",
            "format": ""
        },
        {
            "name": "recentTransfers",
            "baseName": "recent_transfers",
            "type": "Array<Transfer>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Treasury.attributeTypeMap;
    }

    public constructor() {
    }
}