// src/core/modules/dao-api/models/Transfer.ts
/**
 * DAO-API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Token } from '../models/Token';
import { HttpFile } from '../http/http';

export class Transfer {
    'transferId'?: string;
    'daoId': string;
    'tokenId': string;
    'fromAddress': string;
    'toAddress': string;
    'amount': number;
    'timestamp'?: Date;
    'token'?: Token;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "transferId",
            "baseName": "transfer_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "daoId",
            "baseName": "dao_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "tokenId",
            "baseName": "token_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "fromAddress",
            "baseName": "from_address",
            "type": "string",
            "format": ""
        },
        {
            "name": "toAddress",
            "baseName": "to_address",
            "type": "string",
            "format": ""
        },
        {
            "name": "amount",
            "baseName": "amount",
            "type": "number",
            "format": ""
        },
        {
            "name": "timestamp",
            "baseName": "timestamp",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "token",
            "baseName": "token",
            "type": "Token",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Transfer.attributeTypeMap;
    }

    public constructor() {
    }
}